Move "secondary" types close to the code using them
