+++ src/Sdk/NewsroomWebhooks/types.ts
@@ -1,9 +1,9 @@
-import { WebhookEvent } from '../../types';
+import { NewsroomWebhookEvent } from '../../types';
 
 export interface NewsroomWebhookCreateRequest {
     name: string;
     url: string;
-    events?: WebhookEvent[];
+    events?: NewsroomWebhookEvent[];
     secret?: string | null;
     is_active?: boolean;
 }
@@ -11,7 +11,7 @@ export interface NewsroomWebhookCreateRequest {
 export interface NewsroomWebhookUpdateRequest {
     name?: string;
     url?: string;
-    events?: WebhookEvent[];
+    events?: NewsroomWebhookEvent[];
     secret?: string | null;
     is_active?: boolean;
 }
