--- src/types/CoverageProvider.ts
@@ -1,11 +1,0 @@
-export enum CoverageProvider {
-    AMMCO = 'ammco',
-    AUXIPRESS = 'auxipress',
-    BELGA = 'belga',
-    FEED = 'feed',
-    KANTAR = 'kantar',
-    KNEWIN = 'knewin',
-    MEDIAWEB = 'mediaweb',
-    MELTWATER = 'meltwater',
-    MANUAL = 'manual',
-}
