--- src/types/CoverageType.ts
@@ -1,7 +1,0 @@
-export enum CoverageType {
-    ONLINE = 'online',
-    SOCIAL = 'social',
-    PRINT = 'print',
-    RADIO = 'radio',
-    TELEVISION = 'television',
-}
