--- src/types/Utm.ts
@@ -1,7 +1,0 @@
-export interface Utm {
-    utm_id?: string;
-    utm_source?: string;
-    utm_campaign?: string;
-    utm_term?: string;
-    utm_content?: string;
-}
