--- src/types/WebhookEvent.ts
@@ -1,8 +1,0 @@
-export enum WebhookEvent {
-    STORY_CREATED = 'story.created',
-    STORY_UPDATED = 'story.updated',
-    STORY_DELETED = 'story.deleted',
-    CATEGORY_CREATED = 'category.created',
-    CATEGORY_UPDATED = 'category.updated',
-    CATEGORY_DELETED = 'category.deleted',
-}
