+++ src/types/index.ts
@@ -6,7 +6,6 @@ export * from './Campaign';
 export * from './Category';
 export * from './Contact';
 export * from './Coverage';
-export * from './CoverageType';
 export * from './CultureRef';
 export * from './DnsConfigurationInstruction';
 export * from './JobState';
@@ -24,7 +23,6 @@ export * from './SenderAddress';
 export * from './SenderDomain';
 export * from './Story';
 export * from './UserRef';
-export * from './WebhookEvent';
 export * from './UnsubscribeReason';
 export * from './EmailBranding';
 export * from './EmailBrandingMode';
